+++ changes.md
@@ -4,7 +4,7 @@
 1. Fix error in vertical text alignment
 2. Many improvements in TPaletteAxis drawing - draw label, avoid too large ticks.
 3. Fix error with col drawing - bin with maximum value got wrong color  
-4. Test for existing jquery and jquery-ui libraries, reuse when exists 
+4. Test for existing jquery.js, jquery-ui.js and d3.js libraries, reuse when exists 
 
 
 ## Changes in 3.4